refactor of acceptable protobuf messages
