+++ test/managedwriter/test_protos/customer_record.json
